Add Population living in cities or not Add Population who speak different languages Fixed Bugs

+++ src/main/java/com/napier/worldpopulation/City.java
@@ -30,7 +30,7 @@ public class City {
     /**
      * Declare City's population
      */
-    public int Population;
+    public Long Population;
 
     /**
      * Declare Country's Code
